migrationファイルの生成

+++ src/database/migrations/2022_10_21_022647_create_employees_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('employee_id');
             $table->string('employee_name');
             $table->string('employee_pass');
-            $table->unsignedBigInteger('company_id');
+            $table->unsignedInteger('company_id');
             $table->foreign('company_id')->references('company_id')->on('companies');
         });
     }
